+++ docs/changelog.rst
@@ -6,6 +6,12 @@ Changelog
 `Unreleased`_
 -------------
 
+Changed
+~~~~~~~
+
+- The default Hypothesis's ``deadline`` setting for tests with ``schema.parametrize`` is set to 500 ms for consistency with
+  the CLI behaviour. `#705`_
+
 `2.3.3`_ - 2020-08-04
 ---------------------
 
@@ -1303,6 +1309,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#705: https://github.com/kiwicom/schemathesis/issues/705
 .. _#692: https://github.com/kiwicom/schemathesis/issues/692
 .. _#684: https://github.com/kiwicom/schemathesis/issues/684
 .. _#675: https://github.com/kiwicom/schemathesis/issues/675